move CommentChanges into static class within GHEventPayload
